save name textbox uses new input system
but non english letters are not accepted since it keeps the
original condition: if it's not in the font it can't be used.

+++ linuxdoom-1.10/hu_stuff.c
@@ -665,7 +665,8 @@ byte HU_dequeueChatChar(void)
     return c;
 }
 
-int encode_single_utf8(int codepoint, unsigned char *buf);
+int encode_single_utf8(long codepoint, unsigned char *buf);
+
 boolean HU_Responder(event_t *ev)
 {
 
@@ -786,7 +787,7 @@ boolean HU_Responder(event_t *ev)
 
     // proper chat, user typed a character
     if (chat_on && ev->type == ev_character) {
-        int c = toupper(ev->data1);
+        long c = toupper(ev->data1);
         if (' ' <= c && c <= '_') { // uppercase range of ascii printchars
             HU_queueChatChar(c);
             HUlib_addCharToTextLine(&w_chat.l, c);
